requirements.txt
google-generativeai>=0.3.1
PyPDF2>=3.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
tenacity>=8.2.0 